Fix some renamed package imports
