+++ App/src/main/java/cc/softwarefactory/lokki/android/utilities/Utils.java
@@ -30,10 +30,9 @@ import android.support.v4.app.NotificationCompat;
 import android.util.Log;
 import android.widget.Toast;
 
-import cc.softwarefactory.lokki.android.LocationService;
 import cc.softwarefactory.lokki.android.MainApplication;
 import cc.softwarefactory.lokki.android.R;
-import cc.softwarefactory.lokki.android.ServerAPI;
+import cc.softwarefactory.lokki.android.services.LocationService;
 
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GooglePlayServicesUtil;
@@ -391,7 +390,7 @@ public class Utils {
     public static void setVisibility (boolean visible, Context context) {
         try {
             MainApplication.visible = visible;
-            ServerAPI.setVisibility(context, visible);
+            ServerApi.setVisibility(context, visible);
             if (visible == false) {
                 LocationService.stop(context);
             } else {
